最终版

+++ HARDWARE/MPU6050/mpuiic.h
@@ -15,8 +15,8 @@
 //////////////////////////////////////////////////////////////////////////////////
  	   		   
 //IO方向设置
-#define MPU_SDA_IN()  {GPIOB->CRL&=0X0FFFFFFF;GPIOB->CRL|=8<<28;}
-#define MPU_SDA_OUT() {GPIOB->CRL&=0X0FFFFFFF;GPIOB->CRL|=3<<28;}
+#define MPU_SDA_IN()  {GPIOB->CRL&=0X0FFFFFFF;GPIOB->CRL|=(u32)8<<28;}
+#define MPU_SDA_OUT() {GPIOB->CRL&=0X0FFFFFFF;GPIOB->CRL|=(u32)3<<28;}
 
 //IO操作函数	 
 #define MPU_IIC_SCL    PBout(6) 		//SCL
